backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
PyMuPDF
Pillow
pytesseract
